PK값 type 수정

+++ src/main/java/com/green/secondproject/header/HeaderController.java
@@ -23,14 +23,14 @@ public class HeaderController {
     @GetMapping("/school_info")
     @Operation(summary = "학생의 00고등학교 0학년 반 표기",
             description = "요구값 : <br> userId - 학생 PK값<br>")
-    public List<SelSchoolInfoVo> SelectSchoolInfo(@RequestParam int userId) {
+    public List<SelSchoolInfoVo> SelectSchoolInfo(@RequestParam Long userId) {
         return service.selSchoolInfo(userId);
     }
 
     @GetMapping("/school_logo")
     @Operation(summary = "학생의 소속 학교 로고 표기",
             description = "요구값 : <br> userId - 학생 PK값<br>")
-    public List<SelSchoolLogoVo> SelectSchoolLogo(@RequestParam int userId) {
+    public List<SelSchoolLogoVo> SelectSchoolLogo(@RequestParam Long userId) {
         return service.selSchoolLogo(userId);
     }
 }
