+++ src/main/java/com/green/secondproject/myPage/model/SelStudentMyPageVo.java
@@ -6,14 +6,14 @@ import java.time.LocalDate;
 
 @Data
 public class SelStudentMyPageVo {
-    private int userId;
+    private Long userId;
     private String stdnm;
     private String email;
     private String pic;
     private LocalDate birth;
     private String phone;
     private String Address;
-    private int classId;
+    private Long classId;
     private String grade;
     private String van;
     private int schoolId;
