+++ src/main/java/com/green/secondproject/teacher/model/SelUnsignedStudentVo.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 
 @Data
 public class SelUnsignedStudentVo {
-    private int classId;
+    private Long classId;
     private int aprYn;
     private String snm;
     private LocalDate birth;
